requirements.txt
streamlit
pandas
matplotlib
openai
python-dotenv
vnstock
scikit-learn
tensorflow
xgboost
numpy
statsmodels